Backend/AuthService/requirements.txt
fastapi>=0.95.0
uvicorn[standard]
pydantic>=2.5.0
sqlalchemy>=2.0.0
passlib[argon2]>=1.7.4
python-jose>=3.3.0
email_validator